Rename client.js to contract_client.js to avoid confusion.
